feat: rename (webauthn) sessions controller to `webauthn_sessions` to avoid conflicts with Rails' sessions controller generated
