+++ lib/generators/test_unit/webauthn_authentication/templates/test/controllers/webauthn_sessions_controller_test.rb
@@ -1,23 +1,23 @@
 require "test_helper"
 
-class SessionsControllerTest < ActionDispatch::IntegrationTest
+class WebauthnSessionsControllerTest < ActionDispatch::IntegrationTest
   test "should initiate registration successfully" do
     User.create!(username: "alice")
 
-    post get_options_session_url, params: { session: { username: "alice" } }
+    post get_options_webauthn_session_url, params: { session: { username: "alice" } }
 
     assert_response :success
   end
 
   test "should return error if creating session with inexisting username" do
-    post get_options_session_url, params: { session: { username: "alice" } }
+    post get_options_webauthn_session_url, params: { session: { username: "alice" } }
 
     assert_response :unprocessable_entity
     assert_equal [ "Username doesn't exist" ], JSON.parse(response.body)["errors"]
   end
 
   test "should return error if creating session with blank username" do
-    post get_options_session_url, params: { session: { username: "" } }
+    post get_options_webauthn_session_url, params: { session: { username: "" } }
 
     assert_response :unprocessable_entity
     assert_equal [ "Username doesn't exist" ], JSON.parse(response.body)["errors"]
