+++ test/generators/webauthn_authentication_generator_test.rb
@@ -20,7 +20,7 @@ class WebauthnAuthenticationGeneratorTest < Rails::Generators::TestCase
     run_generator_instance
 
     assert_file "app/controllers/registrations_controller.rb"
-    assert_file "app/controllers/sessions_controller.rb"
+    assert_file "app/controllers/webauthn_sessions_controller.rb"
     assert_file "app/controllers/webauthn_credentials_controller.rb"
     assert_file "app/controllers/concerns/authentication.rb"
 
@@ -28,14 +28,14 @@ class WebauthnAuthenticationGeneratorTest < Rails::Generators::TestCase
 
     assert_file "app/views/webauthn_credentials/new.html.erb"
     assert_file "app/views/registrations/new.html.erb"
-    assert_file "app/views/sessions/new.html.erb"
+    assert_file "app/views/webauthn_sessions/new.html.erb"
 
     assert_file "app/javascript/controllers/webauthn_credentials_controller.js"
 
     assert_file "config/initializers/webauthn.rb", /WebAuthn.configure/
 
     assert_file "test/controllers/registrations_controller_test.rb"
-    assert_file "test/controllers/sessions_controller_test.rb"
+    assert_file "test/controllers/webauthn_sessions_controller_test.rb"
     assert_file "test/system/add_credential_test.rb"
     assert_file "test/system/registration_test.rb"
     assert_file "test/system/sign_in_test.rb"
@@ -63,7 +63,7 @@ class WebauthnAuthenticationGeneratorTest < Rails::Generators::TestCase
     run_generator_instance
 
     assert_file "app/controllers/registrations_controller.rb"
-    assert_file "app/controllers/sessions_controller.rb"
+    assert_file "app/controllers/webauthn_sessions_controller.rb"
     assert_file "app/controllers/webauthn_credentials_controller.rb"
     assert_file "app/controllers/concerns/authentication.rb"
 
@@ -71,14 +71,14 @@ class WebauthnAuthenticationGeneratorTest < Rails::Generators::TestCase
 
     assert_file "app/views/webauthn_credentials/new.html.erb"
     assert_file "app/views/registrations/new.html.erb"
-    assert_file "app/views/sessions/new.html.erb"
+    assert_file "app/views/webauthn_sessions/new.html.erb"
 
     assert_file "app/javascript/controllers/webauthn_credentials_controller.js"
 
     assert_file "config/initializers/webauthn.rb", /WebAuthn.configure/
 
     assert_file "test/controllers/registrations_controller_test.rb"
-    assert_file "test/controllers/sessions_controller_test.rb"
+    assert_file "test/controllers/webauthn_sessions_controller_test.rb"
     assert_file "test/system/add_credential_test.rb"
     assert_file "test/system/registration_test.rb"
     assert_file "test/system/sign_in_test.rb"
@@ -105,7 +105,7 @@ class WebauthnAuthenticationGeneratorTest < Rails::Generators::TestCase
     run_generator_instance
 
     assert_file "app/controllers/registrations_controller.rb"
-    assert_file "app/controllers/sessions_controller.rb"
+    assert_file "app/controllers/webauthn_sessions_controller.rb"
     assert_file "app/controllers/webauthn_credentials_controller.rb"
     assert_file "app/controllers/concerns/authentication.rb"
 
@@ -113,7 +113,7 @@ class WebauthnAuthenticationGeneratorTest < Rails::Generators::TestCase
 
     assert_no_file "app/views/webauthn_credentials/new.html.erb"
     assert_no_file "app/views/registrations/new.html.erb"
-    assert_no_file "app/views/sessions/new.html.erb"
+    assert_no_file "app/views/webauthn_sessions/new.html.erb"
 
     assert_file "app/javascript/controllers/webauthn_credentials_controller.js"
